+++ data/Univus.txt
@@ -1,2 +1,2 @@
-[T][ ] return book
 [D][ ] return book (by:Dec 05 2019)
+[E][ ] project meeting (from: Mon 2pm to: 4pm)
