+++ src/main/java/Ui.java
@@ -1,0 +1,22 @@
+import java.util.Scanner;
+
+
+public class Ui {
+    private Scanner scanner;
+    public Ui() {
+        this.scanner = new Scanner(System.in);
+    }
+    public void greet() {
+        System.out.println("____________________________________________________________");
+        System.out.println("Hello! I'm Univus");
+        System.out.println("What can I do for you?");
+        System.out.println("____________________________________________________________");
+    }
+    public void bye() {
+        System.out.println("____________________________________________________________");
+        System.out.println("Bye. Hope to see you again soon!");
+        System.out.println("____________________________________________________________");
+    }
+
+
+}
